+++ CHANGES.md
@@ -1,5 +1,17 @@
 # Change Log
 
+## v0.8.3
+
+Added new actions:
+* `vsphere.custom_attr_assign` - Assign a custom attribute to a given object
+* `vsphere.tags_attach_to_object` - Attach a list of tags to a given object
+
+Contributed by John Schoewe (Encore Technologies).
+
+## v0.8.2
+
+- guest_process_run was missing the workflow file
+
 ## V0.8.2
 
 Added new action: